app/backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pplan-backend"
version = "0.1.0"
description = "PPLAN FastAPI backend"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116,<1.0",
  "uvicorn[standard]>=0.34,<1.0",
  "sqlalchemy>=2.0,<3.0",
  "psycopg[binary]>=3.2,<4.0",
  "alembic>=1.16,<2.0",
  "pydantic-settings>=2.7,<3.0",
  "openpyxl>=3.1,<4.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<9.0",
  "httpx>=0.28,<0.29",
  "ruff>=0.9,<1.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

